+++ waltid-services/waltid-e2e-tests/build.gradle.kts
@@ -26,6 +26,11 @@ dependencies {
     implementation(project(":waltid-services:waltid-wallet-api"))
 
     testImplementation("app.softwork:kotlinx-uuid-core:0.0.26")
+    testImplementation("com.nimbusds:nimbus-jose-jwt:9.37.3")
+    testImplementation("com.augustcellars.cose:cose-java:1.1.0")
+    // Multiplatform / Hashes
+    testImplementation(project.dependencies.platform("org.kotlincrypto.hash:bom:0.4.0"))
+    testImplementation("org.kotlincrypto.hash:sha2")
 
 }
