+++ waltid-services/waltid-issuer-api/src/test/kotlin/id/walt/LspPotentialTest.kt
@@ -64,205 +64,7 @@ import kotlin.io.encoding.ExperimentalEncodingApi
 import kotlin.test.*
 
 object LspPotentialTest {
-  val http = HttpClient {
-    install(ContentNegotiation) {
-      json(Json { explicitNulls = false })
-    }
-    followRedirects = false
-  }
-  var webConfig: WebConfig? = null
 
-  init {
-    runBlocking {
-      WaltidServices.minimalInit()
-    }
-    ConfigManager.testWithConfigs(testConfigs)
-    webConfig = ConfigManager.getConfig<WebConfig>()
-    embeddedServer(
-      CIO,
-      port = webConfig!!.webPort,
-      host = webConfig!!.webHost,
-      configure = {
-      },
-      module = Application::issuerModule
-    ).start(wait = false)
-  }
-
-  @OptIn(ExperimentalEncodingApi::class)
-  @Test
-  fun testLspPotentialTrack1(): Unit = runBlocking {
-    // ### steps 1-6
-    val offerResp = http.get("http://${webConfig!!.webHost}:${webConfig!!.webPort}/openid4vc/lspPotentialCredentialOffer")
-    //val offerResp = http.get("https://issuer.potential.walt-test.cloud/openid4vc/lspPotentialCredentialOffer")
-    assertEquals(HttpStatusCode.OK, offerResp.status)
-
-    val offerUri = offerResp.bodyAsText()
-    //val offerUri = "openid-credential-offer://?credential_offer=%7B%22credential_issuer%22:%22https://mdlpilot.japaneast.cloudapp.azure.com:8017%22,%22credential_configuration_ids%22:%5B%22org.iso.18013.5.1.mDL%22%5D,%22grants%22:%7B%22authorization_code%22:%7B%22issuer_state%22:%22fX35ofcFrYUh2ltjnDZtNXIW7UTUrme767eqPvNix44%22,%22authorization_server%22:%22https://mdlpilot.japaneast.cloudapp.azure.com:8017%22%7D%7D%7D"
-    //val offerUri = "openid-credential-offer://?credential_offer=%7B%22credential_issuer%22%3A%22https%3A%2F%2Flaunchpad.vii.proton.mattrlabs.io%22%2C%22credential_configuration_ids%22%3A%5B%22b59d6a4c-b331-476b-9a4e-ea234c7882c8%22%5D%7D"
-
-    // -------- WALLET ----------
-    // as WALLET: receive credential offer, either being called via deeplink or by scanning QR code
-    // parse credential URI
-    val parsedOffer = OpenID4VCI.parseAndResolveCredentialOfferRequestUrl(offerUri)
-    //assertContains(parsedOffer.credentialConfigurationIds, "potential.light.profile")
-
-    // ### get issuer metadata, steps 7-10
-    val providerMetadataUri = OpenID4VCI.getCIProviderMetadataUrl(parsedOffer.credentialIssuer)
-    val oauthMetadataUri = OpenID4VCI.getOAuthProviderMetadataUrl(parsedOffer.credentialIssuer)
-    val providerMetadata = http.get(providerMetadataUri).bodyAsText().let { OpenIDProviderMetadata.fromJSONString(it) }
-    val oauthMetadata = http.get(oauthMetadataUri).body<OpenIDProviderMetadata>()
-    assertNotNull(providerMetadata.credentialConfigurationsSupported)
-    assertNotNull(providerMetadata.credentialEndpoint)
-    assertNotNull(oauthMetadata.authorizationEndpoint)
-    assertNotNull(oauthMetadata.tokenEndpoint)
-    //assertContains(providerMetadata.codeChallengeMethodsSupported.orEmpty(), "S256")
-
-    // resolve offered credentials
-    val offeredCredentials = OpenID4VCI.resolveOfferedCredentials(parsedOffer, providerMetadata)
-    val offeredCredential = offeredCredentials.first()
-    assertEquals(CredentialFormat.mso_mdoc, offeredCredential.format)
-    assertEquals("org.iso.18013.5.1.mDL", offeredCredential.docType)
-
-    // ### step 11: confirm issuance (nothing to do)
-
-    // ### step 12-15: authorization
-    val codeVerifier = randomUUID()
-
-    val codeChallenge =
-      codeVerifier.let { Base64.UrlSafe.encode(SHA256().digest(it.toByteArray(Charsets.UTF_8))).trimEnd('=') }
-
-    val authReq = AuthorizationRequest(
-      responseType = setOf(ResponseType.Code),
-      clientId = "test-wallet",
-      redirectUri = "https://test-wallet.org",
-      scope = setOf("openid"),
-      issuerState = parsedOffer.grants[GrantType.authorization_code.value]?.issuerState,
-      authorizationDetails = offeredCredentials.map {
-        AuthorizationDetails.fromOfferedCredential(
-          it,
-          providerMetadata.credentialIssuer
-        )
-      },
-      codeChallenge = codeChallenge,
-      codeChallengeMethod = "S256"
-    )
-    val location = if(oauthMetadata.requirePushedAuthorizationRequests != true) {
-      val authResp = http.get(oauthMetadata.authorizationEndpoint!!) {
-        authReq.toHttpParameters().forEach {
-          parameter(it.key, it.value.first())
-        }
-      }
-      assertEquals(HttpStatusCode.Found, authResp.status)
-      Url(authResp.headers[HttpHeaders.Location]!!)
-
-    } else {
-      // TODO: check PAR support, also check issuer of panasonic (https://mdlpilot.japaneast.cloudapp.azure.com:8017), which uses PAR and doesn't seem to work
-      val parResp = http.submitForm(oauthMetadata.pushedAuthorizationRequestEndpoint!!, parametersOf(
-        authReq.toHttpParameters()
-      ))
-      val authResp = http.get(oauthMetadata.authorizationEndpoint!!) {
-        parameter("request_uri", Json.parseToJsonElement(parResp.bodyAsText()).jsonObject["request_uri"]!!.jsonPrimitive.content)
-      }
-      assertEquals(HttpStatusCode.Found, authResp.status)
-      Url(authResp.headers[HttpHeaders.Location]!!)
-    }
-    assertContains(location.parameters.names(), "code")
-    // ### step 16-18: access token retrieval
-    val tokenResp = http.submitForm(oauthMetadata.tokenEndpoint!!,
-      parametersOf(
-        TokenRequest(
-          GrantType.authorization_code, authReq.clientId,
-          authReq.redirectUri, location.parameters["code"]!!,
-          codeVerifier = codeVerifier
-      ).toHttpParameters())).let { TokenResponse.fromJSON(it.body<JsonObject>()) }
-    assertNotNull(tokenResp.accessToken)
-    assertTrue(tokenResp.isSuccess)
-
-    val deviceKeyId = "device-key"
-    val deviceKeyPair: ECKey = runBlocking { ECKey.parseFromPEMEncodedObjects(FileReader("./src/test/resources/ec_device_key.pem").readText()).toECKey() }
-    val certificateChain: String = FileReader("./src/test/resources/cert-device-potential.pem").readText()
-    // ### steps 19-22: credential issuance
-
-    repeat(3) { it ->
-      // TODO: move COSE signing functionality to crypto lib?
-      val coseKey = OneKey(deviceKeyPair.toECPublicKey(), null).AsCBOR().EncodeToBytes()
-      val credReq = CredentialRequest.forOfferedCredential(
-        offeredCredential, ProofOfPossession.CWTProofBuilder(
-          issuerUrl = parsedOffer.credentialIssuer, clientId = authReq.clientId, nonce = tokenResp.cNonce,
-          coseKeyAlgorithm = COSE.AlgorithmID.ECDSA_256.AsCBOR().toString(),
-          coseKey = when(it) { // test with COSE_Key header
-            0 -> coseKey
-            else -> null
-          },
-          x5Cert = when(it) { // test with single x509 certificate in x5chain header
-            1 -> certificateChain.encodeToByteArray()
-            else -> null
-          },
-          x5Chain = when(it) { // test with x509 certificate list in x5chain header
-            2 -> listOf(certificateChain.encodeToByteArray())
-            else -> null
-          }
-        ).build(
-          SimpleCOSECryptoProvider(
-            listOf(
-              COSECryptoProviderKeyInfo(
-                deviceKeyId, AlgorithmID.ECDSA_256,
-                deviceKeyPair.toECPublicKey(), deviceKeyPair.toECPrivateKey()
-              )
-            )
-          ), deviceKeyId
-        )
-      )
-
-      val cwt = Cbor.decodeFromByteArray(COSESign1.serializer(), credReq.proof!!.cwt!!.base64UrlDecode())
-      assertNotNull(cwt.payload)
-      val cwtPayload = Cbor.decodeFromByteArray<MapElement>(cwt.payload!!)
-      assertEquals(DEType.textString, cwtPayload.value.get(MapKey(ProofOfPossession.CWTProofBuilder.LABEL_ISS))?.type)
-      val cwtProtectedHeader = Cbor.decodeFromByteArray<MapElement>(cwt.protectedHeader)
-      assertEquals(cwtProtectedHeader.value[MapKey(ProofOfPossession.CWTProofBuilder.HEADER_LABEL_ALG)]!!.value, -7L)
-      assertEquals(
-        cwtProtectedHeader.value[MapKey(ProofOfPossession.CWTProofBuilder.HEADER_LABEL_CONTENT_TYPE)]!!.value,
-        "openid4vci-proof+cwt"
-      )
-
-      val credResp = http.post(providerMetadata.credentialEndpoint!!) {
-        contentType(ContentType.Application.Json)
-        bearerAuth(tokenResp.accessToken!!)
-        setBody(credReq.toJSON())
-      }.body<JsonObject>().let { CredentialResponse.fromJSON(it) }
-      assertTrue(credResp.isSuccess)
-      assertContains(credResp.customParameters.keys, "credential_encoding")
-      assertEquals("issuer-signed", credResp.customParameters["credential_encoding"]!!.jsonPrimitive.content)
-      assertNotNull(credResp.credential)
-      val mdoc = MDoc(
-        credReq.docType!!.toDE(), IssuerSigned.fromMapElement(
-          Cbor.decodeFromByteArray(credResp.credential!!.jsonPrimitive.content.base64UrlDecode())
-        ), null
-      )
-      assertEquals(credReq.docType, mdoc.docType.value)
-      assertNotNull(mdoc.issuerSigned)
-      assertTrue(
-        mdoc.verifySignature(
-          SimpleCOSECryptoProvider(
-            listOf(
-              LspPotentialInteropEvent.loadPotentialIssuerKeys()
-            )
-          ), LspPotentialInteropEvent.POTENTIAL_ISSUER_KEY_ID
-        )
-      )
-      assertNotNull(mdoc.issuerSigned.nameSpaces)
-      assertNotEquals(0, mdoc.issuerSigned.nameSpaces!!.size)
-      assertNotNull(mdoc.issuerSigned.issuerAuth?.x5Chain)
-      assertTrue(
-        mdoc.verify(
-          MDocVerificationParams(
-            VerificationType.forIssuance, LspPotentialInteropEvent.POTENTIAL_ISSUER_KEY_ID, deviceKeyId,
-            mDocRequest = MDocRequestBuilder(offeredCredential.docType!!).build()
-          ), SimpleCOSECryptoProvider(listOf(LspPotentialInteropEvent.POTENTIAL_ISSUER_CRYPTO_PROVIDER_INFO))
-        )
-      )
-    }
-  }
 
   @Test
   fun testParseCWTExample() {
@@ -275,120 +77,6 @@ object LspPotentialTest {
     val parsedCwt = Cbor.decodeFromHexString(COSESign1.serializer(), data)
     parsedCwt != null
 
-  }
-
-  @OptIn(ExperimentalEncodingApi::class)
-  @Test
-  fun testLspPotentialTrack2(): Unit = runBlocking {
-    // ### steps 1-6
-    val offerResp = http.get("http://${webConfig!!.webHost}:${webConfig!!.webPort}/openid4vc/lspPotentialCredentialOfferT2")
-    //val offerResp = http.get("https://issuer.potential.walt-test.cloud/openid4vc/lspPotentialCredentialOfferT2")
-    assertEquals(HttpStatusCode.OK, offerResp.status)
-
-    val offerUri = offerResp.bodyAsText()
-
-    // -------- WALLET ----------
-    // as WALLET: receive credential offer, either being called via deeplink or by scanning QR code
-    // parse credential URI
-    val parsedOffer = OpenID4VCI.parseAndResolveCredentialOfferRequestUrl(offerUri)
-    //assertContains(parsedOffer.credentialConfigurationIds, "potential.light.profile")
-
-    // ### get issuer metadata, steps 7-10
-    val providerMetadataUri = OpenID4VCI.getCIProviderMetadataUrl(parsedOffer.credentialIssuer)
-    val jwtIssuerMetadataUri = OpenID4VCI.getJWTIssuerProviderMetadataUrl(parsedOffer.credentialIssuer)
-    val providerMetadata = http.get(providerMetadataUri).bodyAsText().let { OpenIDProviderMetadata.fromJSONString(it) }
-    val oauthMetadata = http.get(jwtIssuerMetadataUri).body<OpenIDProviderMetadata>()
-    val jwtIssuerMetadata = http.get(jwtIssuerMetadataUri).body<OpenIDProviderMetadata>()
-    assertNotNull(providerMetadata.credentialConfigurationsSupported)
-    assertNotNull(providerMetadata.credentialEndpoint)
-    assertNotNull(jwtIssuerMetadata.issuer)
-    assertNotNull(jwtIssuerMetadata.jwksUri)
-    assertContains(providerMetadata.codeChallengeMethodsSupported.orEmpty(), "S256")
-
-    val jwks = http.get(jwtIssuerMetadata.jwksUri!!).body<JsonObject>()
-    assertContains(jwks.keys, "keys")
-
-    // resolve offered credentials
-    val offeredCredentials = OpenID4VCI.resolveOfferedCredentials(parsedOffer, providerMetadata)
-    val offeredCredential = offeredCredentials.first()
-    assertEquals(CredentialFormat.sd_jwt_vc, offeredCredential.format)
-    assertEquals("urn:eu.europa.ec.eudi:pid:1", offeredCredential.docType)
-
-    // ### step 11: confirm issuance (nothing to do)
-
-    // ### step 12-15: authorization
-    val codeVerifier = randomUUID()
-
-    val codeChallenge =
-      codeVerifier.let { Base64.UrlSafe.encode(SHA256().digest(it.toByteArray(Charsets.UTF_8))).trimEnd('=') }
-
-    val authReq = AuthorizationRequest(
-      responseType = setOf(ResponseType.Code),
-      clientId = "test-wallet",
-      redirectUri = "https://test-wallet.org",
-      scope = setOf("openid"),
-      issuerState = parsedOffer.grants[GrantType.authorization_code.value]?.issuerState,
-      authorizationDetails = offeredCredentials.map {
-        AuthorizationDetails.fromOfferedCredential(
-          it,
-          providerMetadata.credentialIssuer
-        )
-      },
-      codeChallenge = codeChallenge,
-      codeChallengeMethod = "S256"
-    )
-    val location = if(oauthMetadata.requirePushedAuthorizationRequests != true) {
-      val authResp = http.get(oauthMetadata.authorizationEndpoint!!) {
-        authReq.toHttpParameters().forEach {
-          parameter(it.key, it.value.first())
-        }
-      }
-      assertEquals(HttpStatusCode.Found, authResp.status)
-      Url(authResp.headers[HttpHeaders.Location]!!)
-
-    } else {
-      // TODO: check PAR support, also check issuer of panasonic (https://mdlpilot.japaneast.cloudapp.azure.com:8017), which uses PAR and doesn't seem to work
-      val parResp = http.submitForm(oauthMetadata.pushedAuthorizationRequestEndpoint!!, parametersOf(
-        authReq.toHttpParameters()
-      ))
-      val authResp = http.get(oauthMetadata.authorizationEndpoint!!) {
-        parameter("request_uri", Json.parseToJsonElement(parResp.bodyAsText()).jsonObject["request_uri"]!!.jsonPrimitive.content)
-      }
-      assertEquals(HttpStatusCode.Found, authResp.status)
-      Url(authResp.headers[HttpHeaders.Location]!!)
-    }
-    assertContains(location.parameters.names(), "code")
-    // ### step 16-18: access token retrieval
-    val tokenResp = http.submitForm(oauthMetadata.tokenEndpoint!!,
-      parametersOf(
-        TokenRequest(
-          GrantType.authorization_code, authReq.clientId,
-          authReq.redirectUri, location.parameters["code"]!!,
-          codeVerifier = codeVerifier
-        ).toHttpParameters())).let { TokenResponse.fromJSON(it.body<JsonObject>()) }
-    assertNotNull(tokenResp.accessToken)
-    assertTrue(tokenResp.isSuccess)
-
-    val kpg = KeyPairGenerator.getInstance("EC")
-    kpg.initialize(256)
-    val deviceKeyPair = KeyManager.createKey(KeyGenerationRequest(keyType = KeyType.secp256r1))
-    // ### steps 19-22: credential issuance
-
-    // TODO: move COSE signing functionality to crypto lib?
-    val credReq = CredentialRequest.forOfferedCredential(offeredCredential, ProofOfPossession.JWTProofBuilder(
-      issuerUrl = parsedOffer.credentialIssuer, clientId = authReq.clientId, nonce = tokenResp.cNonce, keyId = null,
-      keyJwk = deviceKeyPair.getPublicKey().exportJWKObject()
-    ).build(deviceKeyPair))
-
-    val credResp = http.post(providerMetadata.credentialEndpoint!!) {
-      contentType(ContentType.Application.Json)
-      bearerAuth(tokenResp.accessToken!!)
-      setBody(credReq.toJSON())
-    }.body<JsonObject>().let { CredentialResponse.fromJSON(it) }
-    assertTrue(credResp.isSuccess)
-    assertNotNull(credResp.credential)
-    val sdJwtVc = SDJwtVC.parse(credResp.credential!!.jsonPrimitive.content)
-    assertNotNull(sdJwtVc.cnfObject)
   }
 
   @OptIn(ExperimentalStdlibApi::class)
